langchain/day4/requirements.txt
langchain
langchain-community
langchain-ollama
python-dotenv
pandas
requests
